+++ tests/functional/transport.cpp
@@ -64,7 +64,8 @@ void transport_data::init_comms() {
 #ifdef CCL_ENABLE_SYCL
     auto sycl_queues = create_sycl_queues("gpu", local_ranks);
     ASSERT(!sycl_queues.empty(), "queues should contain at least one queue");
-    ASSERT(ranks_per_proc == sycl_queues.size(), "ranks and queues sizes should match");
+    ASSERT(static_cast<size_t>(ranks_per_proc) == sycl_queues.size(),
+           "ranks and queues sizes should match");
 
     auto sycl_context = sycl_queues[0].get_context();
     context = ccl::create_context(sycl_context);
